refactor(ai-worker): alinhar estrutura da etapa copy com wireframe
