+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/monitor/CopyPendingJobsService.java
@@ -1,7 +1,7 @@
 package com.marketinghub.worker.geralanding.copy.monitor;
 
 import com.marketinghub.worker.geralanding.copy.dto.GeraLandingStageExecutionDetailDto;
-import com.marketinghub.worker.geralanding.copy.GeraLandingCopyBackendClient;
+import com.marketinghub.worker.geralanding.copy.backend.GeraLandingCopyBackendClient;
 import java.util.List;
 import java.util.Locale;
 import org.springframework.stereotype.Service;
